requirements.txt
blinker==1.9.0
Brotli==1.1.0
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cssselect2==0.8.0
Flask==3.1.1
flask-cors==6.0.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.5
greenlet==3.2.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
qrcode==8.2
reportlab==4.4.2
six==1.17.0
SQLAlchemy==2.0.41
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
